requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
sqlalchemy==2.0.30
pydantic==2.7.1
python-decouple==3.8
loguru==0.7.2
httpx==0.27.0
beautifulsoup4==4.12.3
yt-dlp==2024.5.27
ffmpeg-python==0.2.0
python-telegram-bot==21.3
pytest==8.2.2
pytest-asyncio==0.23.7
